jam: do not set OPTIM by default
If setting OPTIM by jam basically makes it impossible for the project that
is using jam to set its own optimization level while still allowing user
to override that in command line. For example in Haiku jam files there is
line like this: "OPTIM ?= -O2", which obviously is meant to set optimization
level to O2 unless user overwrites it. Unfortunately, because Jam have already
set OPTIM to -O this line never had an intended effect.

After this Jam no longer thinks it knows best what is good for projects
that are using it and Haiku is by default compiled with -O2.

Signed-off-by: Paweł Dziepak <pdziepak@quarnos.org>